Add support for authenticated Maxmind downloads
As at 30-Dec-2019 MaxMind dropped support for unauthenticated
downloads. This PR adds support for the new authenticated
process - which basically involves checking for the presence
of the MAXMIND_LICENSE_KEY env var (stored as a file during
the Buildpack compile step), and then using that license key
as a URL parameter.